+++ compiler/visitor.h
@@ -3,7 +3,7 @@
 
 #include "lexer.h"
 
-void visitor_literal_int(int val);
+void visitor_literal_int(long int val);
 
 void  visitor_op_unary(TokenTyp operator_type);
 void visitor_op_binary(TokenTyp operator_type);
